- Added capacity to update ministrysafe packages
